+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/MedeaFlutterWebrtcPlugin.kt
@@ -1,5 +1,7 @@
 package com.instrumentisto.medea_flutter_webrtc
 
+import android.util.Log
+import com.instrumentisto.medea_flutter_webrtc.controller.ControllerRegistry
 import com.instrumentisto.medea_flutter_webrtc.controller.MediaDevicesController
 import com.instrumentisto.medea_flutter_webrtc.controller.PeerConnectionFactoryController
 import com.instrumentisto.medea_flutter_webrtc.controller.VideoRendererFactoryController
@@ -9,6 +11,8 @@ import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.view.TextureRegistry
 
+private val TAG = MedeaFlutterWebrtcPlugin::class.java.simpleName
+
 class MedeaFlutterWebrtcPlugin : FlutterPlugin, ActivityAware {
   private var peerConnectionFactory: PeerConnectionFactoryController? = null
   private var mediaDevices: MediaDevicesController? = null
@@ -16,19 +20,39 @@ class MedeaFlutterWebrtcPlugin : FlutterPlugin, ActivityAware {
   private var messenger: BinaryMessenger? = null
   private var state: State? = null
   private var textureRegistry: TextureRegistry? = null
+  private var permissions: Permissions? = null
+  private var activityPluginBinding: ActivityPluginBinding? = null
 
   override fun onAttachedToEngine(registrar: FlutterPlugin.FlutterPluginBinding) {
+    Log.i(TAG, "Attached to engine")
+
     messenger = registrar.binaryMessenger
     state = State(registrar.applicationContext)
     textureRegistry = registrar.textureRegistry
   }
 
-  override fun onDetachedFromEngine(registrar: FlutterPlugin.FlutterPluginBinding) {}
+  override fun onDetachedFromEngine(registrar: FlutterPlugin.FlutterPluginBinding) {
+    Log.i(TAG, "Detached from engine")
+
+    ControllerRegistry.disposeAll()
+
+    messenger = null
+    state = null
+    textureRegistry = null
+    activityPluginBinding = null
+    permissions = null
+    mediaDevices = null
+    peerConnectionFactory = null
+    videoRendererFactory = null
+  }
 
   override fun onAttachedToActivity(registrar: ActivityPluginBinding) {
-    val permissions = Permissions(registrar.activity)
-    registrar.addRequestPermissionsResultListener(permissions)
-    mediaDevices = MediaDevicesController(messenger!!, state!!, permissions)
+    Log.i(TAG, "Attached to activity")
+
+    activityPluginBinding = registrar
+    permissions = Permissions(activityPluginBinding!!.activity)
+    activityPluginBinding!!.addRequestPermissionsResultListener(permissions!!)
+    mediaDevices = MediaDevicesController(messenger!!, state!!, permissions!!)
     peerConnectionFactory = PeerConnectionFactoryController(messenger!!, state!!)
     videoRendererFactory = VideoRendererFactoryController(messenger!!, textureRegistry!!)
   }
@@ -37,5 +61,10 @@ class MedeaFlutterWebrtcPlugin : FlutterPlugin, ActivityAware {
 
   override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {}
 
-  override fun onDetachedFromActivity() {}
+  override fun onDetachedFromActivity() {
+    Log.i(TAG, "Detached from activity")
+
+    activityPluginBinding!!.removeRequestPermissionsResultListener(permissions!!)
+    activityPluginBinding = null
+  }
 }