+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/State.kt
@@ -15,10 +15,7 @@ import org.webrtc.audio.JavaAudioDeviceModule
  *
  * @property context Android [Context] used, for example, for `getUserMedia` requests.
  */
-class State(private val context: Context) {
-  /** Module for the controlling audio devices in context of `libwebrtc`. */
-  private var audioDeviceModule: JavaAudioDeviceModule? = null
-
+class State(val context: Context) {
   /** [VideoEncoderFactory] used by the [PeerConnectionFactory]. */
   var encoder: WebrtcVideoEncoderFactory
 
@@ -50,8 +47,8 @@ class State(private val context: Context) {
    * @return Current [PeerConnectionFactory] of this [State].
    */
   fun getPeerConnectionFactory(): PeerConnectionFactory {
-    if (factory == null) {
-      audioDeviceModule =
+    if (factory == null || factory!!.nativeOwnedFactoryAndThreads == 0L) {
+      var audioDeviceModule =
           JavaAudioDeviceModule.builder(context)
               .setUseHardwareAcousticEchoCanceler(true)
               .setUseHardwareNoiseSuppressor(true)
@@ -64,16 +61,10 @@ class State(private val context: Context) {
               .setVideoDecoderFactory(decoder)
               .setAudioDeviceModule(audioDeviceModule)
               .createPeerConnectionFactory()
-
-      audioDeviceModule!!.setSpeakerMute(false)
+      audioDeviceModule.release()
     }
 
     return factory!!
-  }
-
-  /** @return Android SDK [Context]. */
-  fun getAppContext(): Context {
-    return context
   }
 
   /** @return [AudioManager] system service. */