+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/controller/PeerConnectionController.kt
@@ -10,8 +10,10 @@ import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.launch
 
 /**
@@ -23,7 +25,10 @@ import kotlinx.coroutines.launch
 class PeerConnectionController(
     private val messenger: BinaryMessenger,
     private val peer: PeerConnectionProxy
-) : MethodChannel.MethodCallHandler, EventChannel.StreamHandler, IdentifiableController {
+) : EventChannel.StreamHandler, Controller {
+  /** [CoroutineScope] for this [PeerConnectionController] */
+  private val scope = CoroutineScope(SupervisorJob() + Dispatchers.Main)
+
   /** Unique ID of the [MethodChannel] of this controller. */
   private val channelId = nextChannelId()
 
@@ -84,12 +89,13 @@ class PeerConnectionController(
     chan.setMethodCallHandler(this)
     eventChannel.setStreamHandler(this)
     peer.addEventObserver(eventObserver)
+    ControllerRegistry.register(this)
   }
 
   override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
     when (call.method) {
       "createOffer" -> {
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             result.success(peer.createOffer().asFlutterResult())
           } catch (e: Exception) {
@@ -98,7 +104,7 @@ class PeerConnectionController(
         }
       }
       "createAnswer" -> {
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             result.success(peer.createAnswer().asFlutterResult())
           } catch (e: Exception) {
@@ -114,7 +120,7 @@ class PeerConnectionController(
             } else {
               SessionDescription.fromMap(descriptionArg)
             }
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             peer.setLocalDescription(description)
             result.success(null)
@@ -125,7 +131,7 @@ class PeerConnectionController(
       }
       "setRemoteDescription" -> {
         val descriptionArg: Map<String, Any> = call.argument("description")!!
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             peer.setRemoteDescription(SessionDescription.fromMap(descriptionArg))
             result.success(null)
@@ -136,7 +142,7 @@ class PeerConnectionController(
       }
       "addIceCandidate" -> {
         val candidate: Map<String, Any> = call.argument("candidate")!!
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             peer.addIceCandidate(IceCandidate.fromMap(candidate))
             result.success(null)
@@ -155,12 +161,9 @@ class PeerConnectionController(
               } else {
                 peer.addTransceiver(mediaType, RtpTransceiverInit.fromMap(transceiverInitArg))
               }
-          if (transceiver != null) {
-            val transceiverController = RtpTransceiverController(messenger, transceiver)
-            result.success(transceiverController.asFlutterResult())
-          } else {
-            result.success(null)
-          }
+
+          val transceiverController = RtpTransceiverController(messenger, transceiver)
+          result.success(transceiverController.asFlutterResult())
         } catch (e: Exception) {
           resultUnhandledException(result, e)
         }
@@ -176,7 +179,7 @@ class PeerConnectionController(
         result.success(null)
       }
       "getStats" -> {
-        GlobalScope.launch(Dispatchers.Main) {
+        scope.launch {
           try {
             val stats = peer.getStats()
             if (stats == null) {
@@ -190,7 +193,7 @@ class PeerConnectionController(
         }
       }
       "dispose" -> {
-        dispose()
+        disposeInternal(false)
         result.success(null)
       }
     }
@@ -217,13 +220,22 @@ class PeerConnectionController(
   fun asFlutterResult(): Map<String, Any> =
       mapOf<String, Any>("channelId" to channelId, "id" to peer.id)
 
+  override fun dispose() {
+    disposeInternal(true)
+  }
+
   /**
    * Closes method and event channels of this [PeerConnectionController].
    *
    * Disposes underlying [PeerConnectionProxy].
    */
-  private fun dispose() {
-    peer.dispose()
+  private fun disposeInternal(cancel: Boolean) {
+    ControllerRegistry.unregister(this)
+    scope.cancel("disposed")
     chan.setMethodCallHandler(null)
+    peer.dispose()
+    if (cancel) {
+      onCancel(null)
+    }
   }
 }