+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/controller/RtpSenderController.kt
@@ -13,7 +13,7 @@ import io.flutter.plugin.common.MethodChannel
  * @property sender Underlying [RtpSenderProxy] to perform [MethodCall]s on.
  */
 class RtpSenderController(messenger: BinaryMessenger, private val sender: RtpSenderProxy) :
-    MethodChannel.MethodCallHandler, IdentifiableController {
+    Controller {
   /** Unique ID of the [MethodChannel] of this controller. */
   private val channelId = nextChannelId()
 
@@ -22,6 +22,7 @@ class RtpSenderController(messenger: BinaryMessenger, private val sender: RtpSen
 
   init {
     chan.setMethodCallHandler(this)
+    ControllerRegistry.register(this)
   }
 
   override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
@@ -82,10 +83,16 @@ class RtpSenderController(messenger: BinaryMessenger, private val sender: RtpSen
         result.success(null)
       }
       "dispose" -> {
-        chan.setMethodCallHandler(null)
+        dispose()
         result.success(null)
       }
     }
+  }
+
+  override fun dispose() {
+    ControllerRegistry.unregister(this)
+    chan.setMethodCallHandler(null)
+    sender.setDisposed()
   }
 
   /**