+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/controller/VideoRendererController.kt
@@ -17,8 +17,8 @@ import io.flutter.plugin.common.MethodChannel
  */
 class VideoRendererController(
     messenger: BinaryMessenger,
-    private val videoRenderer: FlutterRtcVideoRenderer
-) : MethodChannel.MethodCallHandler, EventChannel.StreamHandler, IdentifiableController {
+    private val videoRenderer: FlutterRtcVideoRenderer,
+) : EventChannel.StreamHandler, Controller {
   /** Unique ID of the [MethodChannel] of this controller. */
   private val channelId: Long = nextChannelId()
 
@@ -36,6 +36,7 @@ class VideoRendererController(
   init {
     chan.setMethodCallHandler(this)
     eventChannel.setStreamHandler(this)
+    ControllerRegistry.register(this)
 
     videoRenderer.setEventListener(
         object : FlutterRtcVideoRenderer.Companion.EventListener {
@@ -84,6 +85,8 @@ class VideoRendererController(
   }
 
   override fun onCancel(obj: Any?) {
+    eventChannel.setStreamHandler(null)
+    eventSink?.endOfStream()
     eventSink = null
   }
 
@@ -95,14 +98,10 @@ class VideoRendererController(
   fun asFlutterResult(): Map<String, Any> =
       mapOf("channelId" to channelId, "textureId" to videoRenderer.textureId())
 
-  /**
-   * Closes method channel of this [VideoRendererController].
-   *
-   * Disposes underlying [FlutterRtcVideoRenderer].
-   */
-  private fun dispose() {
-    eventChannel.setStreamHandler(null)
-    eventSink = null
+  override fun dispose() {
+    ControllerRegistry.unregister(this)
+    chan.setMethodCallHandler(null)
+    onCancel(null)
     videoRenderer.dispose()
   }
 }