+++ android/src/main/kotlin/com/instrumentisto/medea_flutter_webrtc/proxy/Proxy.kt
@@ -9,8 +9,7 @@ import org.webrtc.ThreadUtils
  * `RtpSender` will be outdated. To keep this from happening `PeerConnection` should update its
  * `RtpSender`s with a newly obtained `RtpSender`s via [Proxy.replace] method.
  */
-abstract class Proxy<T>
-constructor(
+abstract class Proxy<T>(
     /** Underlying `libwebrtc` object of this proxy. */
     var obj: T
 ) {