+++ CHANGELOG.md
@@ -16,12 +16,17 @@ All user visible changes to this project will be documented in this file. This p
 
 ### Changed
 
-- Upgraded [libwebrtc] to [137.0.7151.68] version. ([#203], [todo])
+- Upgraded [libwebrtc] to [137.0.7151.68] version. ([#203], [007bd441])
 - Upgraded [`flutter_rust_bridge`] crate to [2.10.0][frb-2.10.0] version. ([#201])
 
+### Fixed
+
+- Resources cleanup when `medea_flutter_webrtc` Flutter plugin is detached on Android. ([#202])
+
 [#201]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/201
+[#202]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/202
 [#203]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/203
-[todo]: https://github.com/instrumentisto/medea-flutter-webrtc/commit/todo
+[007bd441]: https://github.com/instrumentisto/medea-flutter-webrtc/commit/007bd441ce5386aa84e47b3dcc49bdfee070241d
 [137.0.7151.68]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/137.0.7151.68
 [frb-2.10.0]: https://github.com/fzyzcjy/flutter_rust_bridge/releases/tag/v2.10.0